# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de Controle de Ponto
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface Funcionario {
  id: number;
  codigo: number;
  nome: string;
  cargo: string;
  vinculo: string;
  situacao: string;
  adiantamento: number;
  transporte: boolean;
  vale_alimentacao?: boolean;
  jornada_diaria: string;
  ativo: boolean;
  empresa?: string | null;
  data_contrato?: string | null;
  salario?: string | null;
  endereco?: string | null;
  numero?: string | null;
  bairro?: string | null;
  cidade?: string | null;
  cep?: string | null;
  estado_civil?: string | null;
  raca_cor?: string | null;
  horario?: string | null;
  escolaridade?: string | null;
  pis?: string | null;
  escala_quinzenal: boolean;
  quinzena_referencia?: string | null;
  he_100_acima_2h: boolean;
  criado_em: string;
}

export type CreateFuncionarioBodyVinculo =
  (typeof CreateFuncionarioBodyVinculo)[keyof typeof CreateFuncionarioBodyVinculo];

export const CreateFuncionarioBodyVinculo = {
  CLT: "CLT",
  Contribuinte: "Contribuinte",
  Autonomo: "Autonomo",
  Estagiario: "Estagiario",
} as const;

export type CreateFuncionarioBodySituacao =
  (typeof CreateFuncionarioBodySituacao)[keyof typeof CreateFuncionarioBodySituacao];

export const CreateFuncionarioBodySituacao = {
  Ativo: "Ativo",
  Demitido: "Demitido",
  Aviso: "Aviso",
  Ferias: "Ferias",
} as const;

export interface CreateFuncionarioBody {
  codigo: number;
  nome: string;
  cargo?: string;
  vinculo?: CreateFuncionarioBodyVinculo;
  situacao?: CreateFuncionarioBodySituacao;
  adiantamento?: number;
  transporte?: boolean;
  vale_alimentacao?: boolean;
  jornada_diaria?: string;
  ativo?: boolean;
  empresa?: string | null;
  data_contrato?: string | null;
  salario?: string | null;
  endereco?: string | null;
  numero?: string | null;
  bairro?: string | null;
  cidade?: string | null;
  cep?: string | null;
  estado_civil?: string | null;
  raca_cor?: string | null;
  horario?: string | null;
  escolaridade?: string | null;
  pis?: string | null;
  escala_quinzenal?: boolean;
  quinzena_referencia?: string | null;
  he_100_acima_2h?: boolean;
}

export interface UpdateFuncionarioBody {
  codigo?: number;
  nome?: string;
  cargo?: string;
  vinculo?: string;
  situacao?: string;
  adiantamento?: number;
  transporte?: boolean;
  vale_alimentacao?: boolean;
  jornada_diaria?: string;
  ativo?: boolean;
  empresa?: string | null;
  data_contrato?: string | null;
  salario?: string | null;
  endereco?: string | null;
  numero?: string | null;
  bairro?: string | null;
  cidade?: string | null;
  cep?: string | null;
  estado_civil?: string | null;
  raca_cor?: string | null;
  horario?: string | null;
  escolaridade?: string | null;
  pis?: string | null;
  escala_quinzenal?: boolean;
  quinzena_referencia?: string | null;
  he_100_acima_2h?: boolean;
}

export interface FuncionarioArquivo {
  id: number;
  funcionario_id: number;
  nome_arquivo: string;
  tipo_arquivo: string;
  caminho: string;
  criado_em: string;
}

export type RegistroPontoJustificativa =
  (typeof RegistroPontoJustificativa)[keyof typeof RegistroPontoJustificativa];

export const RegistroPontoJustificativa = {
  nenhuma: "nenhuma",
  justificada: "justificada",
  injustificada: "injustificada",
} as const;

export type RegistroPontoTipoDia =
  (typeof RegistroPontoTipoDia)[keyof typeof RegistroPontoTipoDia];

export const RegistroPontoTipoDia = {
  normal: "normal",
  feriado: "feriado",
  feriado_trabalhado: "feriado_trabalhado",
  falta: "falta",
  falta_justificada: "falta_justificada",
  atraso_justificado: "atraso_justificado",
} as const;

export interface RegistroPonto {
  id: number;
  funcionario_id: number;
  data: string;
  entrada?: string | null;
  saida?: string | null;
  saida_almoco?: string | null;
  volta_almoco?: string | null;
  intervalo?: string | null;
  total_horas?: string | null;
  he_60?: string | null;
  he_100?: string | null;
  atrasos?: string | null;
  faltas?: string | null;
  observacoes?: string | null;
  justificativa?: RegistroPontoJustificativa;
  horas_justificadas?: string | null;
  tipo_dia: RegistroPontoTipoDia;
}

export type UpsertRegistroBodyJustificativa =
  (typeof UpsertRegistroBodyJustificativa)[keyof typeof UpsertRegistroBodyJustificativa];

export const UpsertRegistroBodyJustificativa = {
  nenhuma: "nenhuma",
  justificada: "justificada",
  injustificada: "injustificada",
} as const;

export type UpsertRegistroBodyTipoDia =
  (typeof UpsertRegistroBodyTipoDia)[keyof typeof UpsertRegistroBodyTipoDia];

export const UpsertRegistroBodyTipoDia = {
  normal: "normal",
  feriado: "feriado",
  feriado_trabalhado: "feriado_trabalhado",
  falta: "falta",
  falta_justificada: "falta_justificada",
  atraso_justificado: "atraso_justificado",
} as const;

export interface UpsertRegistroBody {
  funcionario_id: number;
  data: string;
  entrada?: string | null;
  saida?: string | null;
  saida_almoco?: string | null;
  volta_almoco?: string | null;
  intervalo?: string | null;
  he_60?: string | null;
  he_100?: string | null;
  atrasos?: string | null;
  faltas?: string | null;
  observacoes?: string | null;
  justificativa?: UpsertRegistroBodyJustificativa;
  tipo_dia?: UpsertRegistroBodyTipoDia;
}

export type BaterPontoBodyTipo =
  (typeof BaterPontoBodyTipo)[keyof typeof BaterPontoBodyTipo];

export const BaterPontoBodyTipo = {
  entrada: "entrada",
  saida_almoco: "saida_almoco",
  volta_almoco: "volta_almoco",
  saida: "saida",
} as const;

export interface BaterPontoBody {
  funcionario_id: number;
  tipo: BaterPontoBodyTipo;
}

export interface BaterPontoResponse {
  funcionario_id: number;
  tipo: string;
  horario: string;
  data: string;
  registro: RegistroPonto;
}

export interface ResumoMes {
  total_horas: string;
  he_60: string;
  he_100: string;
  atrasos: string;
  faltas_dia: number;
  faltas_horas: string;
  dias_trabalhados: number;
  dom_feriados: number;
  horas_justificadas: string;
  dias_justificados: number;
}

export interface FolhaMensal {
  funcionario: Funcionario;
  mes: string;
  registros: RegistroPonto[];
  resumo: ResumoMes;
}

export interface ConsolidadoLinha {
  funcionario_id: number;
  codigo?: number | null;
  nome: string;
  total_horas: string;
  he_60: string;
  he_100: string;
  atrasos: string;
  faltas: number;
  dias_trabalhados: number;
  dom_feriados: number;
  horas_justificadas: string;
  dias_justificados: number;
  adiantamento: number;
}

export interface Consolidado {
  mes: string;
  linhas: ConsolidadoLinha[];
  total_geral: ConsolidadoLinha;
}

export interface ResumoFuncionario {
  id: number;
  codigo: number;
  nome: string;
  cargo: string;
  vinculo: string;
  situacao: string;
  adiantamento: number;
  transporte: boolean;
  vale_alimentacao?: boolean;
  jornada_diaria: string;
  faltas_dia: number;
  faltas_horas: string;
  he_60: string;
  he_100: string;
  horas_justificadas: string;
  dias_justificados: number;
}

export interface ImportacaoResponse {
  importados: number;
  erros: string[];
  mes: string;
}

export type GetFuncionariosParams = {
  situacao?: string;
  vinculo?: string;
  ativo?: boolean;
};

export type UploadFuncionarioArquivoBody = {
  file: Blob;
};

export type GetRegistrosFuncionarioParams = {
  mes: string;
};

export type GetConsolidadoParams = {
  mes: string;
};

export type GetResumoParams = {
  mes: string;
  situacao?: string;
  vinculo?: string;
};

export type ExportarFolhaParams = {
  mes: string;
};

export type ImportarExcelParams = {
  funcionario_id: number;
  mes: string;
};