# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API de Controle de Ponto
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /funcionarios:
    get:
      operationId: getFuncionarios
      tags: [funcionarios]
      summary: Listar funcionários
      parameters:
        - name: situacao
          in: query
          schema:
            type: string
        - name: vinculo
          in: query
          schema:
            type: string
        - name: ativo
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de funcionários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Funcionario"
    post:
      operationId: createFuncionario
      tags: [funcionarios]
      summary: Criar funcionário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFuncionarioBody"
      responses:
        "201":
          description: Funcionário criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
  /funcionarios/{id}:
    get:
      operationId: getFuncionario
      tags: [funcionarios]
      summary: Buscar funcionário por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Funcionário encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateFuncionario
      tags: [funcionarios]
      summary: Atualizar funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFuncionarioBody"
      responses:
        "200":
          description: Funcionário atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funcionario"
    delete:
      operationId: deleteFuncionario
      tags: [funcionarios]
      summary: Remover funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /funcionarios/{id}/arquivos:
    get:
      operationId: getFuncionarioArquivos
      tags: [funcionarios]
      summary: Listar arquivos do funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de arquivos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FuncionarioArquivo"
    post:
      operationId: uploadFuncionarioArquivo
      tags: [funcionarios]
      summary: Enviar arquivo do funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Arquivo enviado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FuncionarioArquivo"
  /funcionarios/{id}/arquivos/{arquivoId}:
    delete:
      operationId: deleteFuncionarioArquivo
      tags: [funcionarios]
      summary: Remover arquivo do funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: arquivoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arquivo removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Funcionario:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: integer
        nome:
          type: string
        cargo:
          type: string
        vinculo:
          type: string
        situacao:
          type: string
        adiantamento:
          type: number
          format: float
        transporte:
          type: boolean
        vale_alimentacao:
          type: boolean
        jornada_diaria:
          type: string
        ativo:
          type: boolean
        empresa:
          type: string
          nullable: true
        data_contrato:
          type: string
          nullable: true
        salario:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        numero:
          type: string
          nullable: true
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        estado_civil:
          type: string
          nullable: true
        raca_cor:
          type: string
          nullable: true
        horario:
          type: string
          nullable: true
        escolaridade:
          type: string
          nullable: true
        pis:
          type: string
          nullable: true
        escala_quinzenal:
          type: boolean
        quinzena_referencia:
          type: string
          nullable: true
        he_100_acima_2h:
          type: boolean
        criado_em:
          type: string
      required:
        - id
        - codigo
        - nome
        - cargo
        - vinculo
        - situacao
        - adiantamento
        - transporte
        - jornada_diaria
        - ativo
        - escala_quinzenal
        - he_100_acima_2h
        - criado_em
    CreateFuncionarioBody:
      type: object
      properties:
        codigo:
          type: integer
        nome:
          type: string
        cargo:
          type: string
        vinculo:
          type: string
          enum: [CLT, Contribuinte, Autonomo, Estagiario]
        situacao:
          type: string
          enum: [Ativo, Demitido, Aviso, Ferias]
        adiantamento:
          type: number
          format: float
        transporte:
          type: boolean
        vale_alimentacao:
          type: boolean
        jornada_diaria:
          type: string
        ativo:
          type: boolean
        empresa:
          type: string
          nullable: true
        data_contrato:
          type: string
          nullable: true
        salario:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        numero:
          type: string
          nullable: true
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        estado_civil:
          type: string
          nullable: true
        raca_cor:
          type: string
          nullable: true
        horario:
          type: string
          nullable: true
        escolaridade:
          type: string
          nullable: true
        pis:
          type: string
          nullable: true
        escala_quinzenal:
          type: boolean
        quinzena_referencia:
          type: string
          nullable: true
        he_100_acima_2h:
          type: boolean
      required:
        - codigo
        - nome
    UpdateFuncionarioBody:
      type: object
      properties:
        codigo:
          type: integer
        nome:
          type: string
        cargo:
          type: string
        vinculo:
          type: string
        situacao:
          type: string
        adiantamento:
          type: number
          format: float
        transporte:
          type: boolean
        vale_alimentacao:
          type: boolean
        jornada_diaria:
          type: string
        ativo:
          type: boolean
        empresa:
          type: string
          nullable: true
        data_contrato:
          type: string
          nullable: true
        salario:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        numero:
          type: string
          nullable: true
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        estado_civil:
          type: string
          nullable: true
        raca_cor:
          type: string
          nullable: true
        horario:
          type: string
          nullable: true
        escolaridade:
          type: string
          nullable: true
        pis:
          type: string
          nullable: true
        escala_quinzenal:
          type: boolean
        quinzena_referencia:
          type: string
          nullable: true
        he_100_acima_2h:
          type: boolean
    FuncionarioArquivo:
      type: object
      properties:
        id:
          type: integer
        funcionario_id:
          type: integer
        nome_arquivo:
          type: string
        tipo_arquivo:
          type: string
        caminho:
          type: string
        criado_em:
          type: string
      required:
        - id
        - funcionario_id
        - nome_arquivo
        - tipo_arquivo
        - caminho
        - criado_em
...